+++ src/_optPlug.node/errors.ts
@@ -17,24 +17,11 @@ SPDX-License-Identifier: Apache-2.0
 Copyright (c) OWASP Foundation. All Rights Reserved.
 */
 
-/**
- * mocha config
- * @see {@link https://mochajs.org/#configuring-mocha-nodejs}
- * @see {@link https://github.com/mochajs/mocha/blob/master/example/config/.mocharc.js example}
- * @type {import('@types/mocha').Mocha.MochaOptions}
- */
-module.exports = {
-  timeout: 10000,
-  spec: [
-    'tests'
-  ],
-  recursive: true,
-  parallel: false, // if true, then some IDEs cannot run it
-  global: [],
-  extension: [
-    'spec.js', 'test.js',
-    'spec.cjs', 'test.cjs',
-    'spec.mjs', 'test.mjs',
-  ],
-  ui: 'tdd',
+export class OptPlugError extends Error {
+  readonly cause: any | undefined
+
+  constructor (message: string, cause?: any) {
+    super(message)
+    this.cause = cause
+  }
 }