+++ src/_optPlug.node/xmlValidator.ts
@@ -17,24 +17,15 @@ SPDX-License-Identifier: Apache-2.0
 Copyright (c) OWASP Foundation. All Rights Reserved.
 */
 
-/**
- * mocha config
- * @see {@link https://mochajs.org/#configuring-mocha-nodejs}
- * @see {@link https://github.com/mochajs/mocha/blob/master/example/config/.mocharc.js example}
- * @type {import('@types/mocha').Mocha.MochaOptions}
- */
-module.exports = {
-  timeout: 10000,
-  spec: [
-    'tests'
-  ],
-  recursive: true,
-  parallel: false, // if true, then some IDEs cannot run it
-  global: [],
-  extension: [
-    'spec.js', 'test.js',
-    'spec.cjs', 'test.cjs',
-    'spec.mjs', 'test.mjs',
-  ],
-  ui: 'tdd',
-}
+import type { ValidationError } from '../validation/types'
+import opWrapper, { type WillThrow } from './_wrapper'
+
+export type Validator = (data: string) => null | ValidationError
+export type Functionality = (schemaPath: string) => Promise<Validator>
+
+export default opWrapper<Functionality>('XmlValidator', [
+  /* eslint-disable @typescript-eslint/no-var-requires */
+  ['libxmljs2', () => require('./__xmlValidators/libxmljs2').default]
+  // ... add others here, pull-requests welcome!
+  /* eslint-enable @typescript-eslint/no-var-requires */
+]) satisfies Functionality | WillThrow