+++ src/serialize/baseSerializer.ts
@@ -58,10 +58,6 @@ export abstract class BaseSerializer<NormalizedBom> implements Serializer {
     }
   }
 
-  /**
-   * @readonly
-   * @throws {@link Error}
-   */
   public serialize (bom: Bom, options?: SerializerOptions & NormalizerOptions): string {
     return this._serialize(
       this.#normalize(bom, options),